requirements.txt
absl-py==1.4.0
amqp==5.1.1
asgiref==3.6.0
attrs==22.2.0
azure-ai-ml==1.4.0
azure-common==1.1.28
azure-core==1.26.3
azure-identity==1.12.0
azure-mgmt-core==1.3.2
azure-storage-blob==12.15.0
azure-storage-file-datalake==12.10.1
azure-storage-file-share==12.11.1
beautifulsoup4==4.11.2
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
Django==4.1.7
django-bootstrap-v5==1.0.11
django-bootstrap5==22.2
django-celery-results==2.5.0
django-storages==1.13.2
dnspython==2.3.0
docker==6.0.1
filelock==3.10.0
fonttools==4.39.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib==3.7.1
mpmath==1.3.0
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pip==22.3.1
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydash==5.1.2
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
setuptools==67.6.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4
sqlparse==0.4.3
strictyaml==1.7.3
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.5.1
Werkzeug==2.2.3
wheel==0.40.0
